Update internal API
