+++ tsconfig-client.json
@@ -7,6 +7,11 @@
         "outDir": "dist/client",
         "rootDir": "sources/lib"
     },
-    "files": [
+    "exclude": [
+        "dist",
+        "node_modules"
+    ],
+    "include": [
+        "sources/lib/**/*.ts"
     ]
 }
