+++ tsconfig.json
@@ -14,11 +14,6 @@
         "node_modules"
     ],
     "include": [
-        "sources/lib/**/*.ts"
-    ],
-    "files": [
-        "sources/bin/ordbok-assembler.ts",
-        "sources/dictionary-plugin/ordbok-plugin.ts",
-        "sources/plugin.ts"
+        "sources/**/*.ts"
     ]
 }
